game_review/requirements.txt
asgiref==3.8.1
Django==5.0.3
sqlparse==0.4.4
asgiref==3.8.1
Django==5.0.3
django-taggit==5.0.1
sqlparse==0.4.4
asgiref==3.8.1
Django==5.0.3
django-bootstrap5==23.4
django-taggit==5.0.1
sqlparse==0.4.4
